Dev (#3)
* Implemented UI of home page

* Hooked up news, weather, and quotes to their respective API

* Added a simple landing page

* Made small UI changes to Home. Added UI for WorkMap

* Hooked up Workmap to firebase backend

* Updated weather background to match current weather

* Updated various UI elements. Replaced Quote section with Upcoming due dates.

* Made various UI updates

* Refactored home components to use ipinfo.io to get location

* Refactored and implemented catches for users not signed in

* Updated Top News to use gnews.io for production reasons. Made small UI updates to various components

+++ src/components/Workmap.jsx
@@ -69,10 +69,12 @@ function Workmap() {
     }
 
     const handleDelete = () => {
-        const itemsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/items");
-        itemsRef.doc(itemId).delete().then(() => {
-            setModalOpen(false);
-        });
+        if (user) {
+            const itemsRef = firebaseApp.firestore().collection("workmap/" + user.uid + "/items");
+            itemsRef.doc(itemId).delete().then(() => {
+                setModalOpen(false);
+            });
+        }
     }
 
     return (
@@ -90,7 +92,7 @@ function Workmap() {
                             itemAbbrev={itemAbbrev} setItemAbbrev={setItemAbbrev}
                             itemDesc={itemDesc} setItemDesc={setItemDesc}
                             itemDue={itemDue} setItemDue={setItemDue} modalNew={modalNew}
-                            onSave={handleSave} onDelete={handleDelete}/>
+                            onSave={handleSave} onDelete={handleDelete} user={user}/>
             </ReactModal>
         </div>
     );
